updating 9-fizz_buzz.c file

+++ 0x04-more_functions_nested_loops/5-more_numbers.c
@@ -4,7 +4,7 @@
  */
 void more_numbers(void)
 {
-	int j, i = 0;
+	long j, i = 0; 
 
 	while (i < 10)
 	{
@@ -15,7 +15,7 @@ void more_numbers(void)
 				_putchar((j / 10) + '0');
 				_putchar((j % 10) + '0');
 			}
-			else
+                        else
 				_putchar(j + '0');
 		}
 		j = 0;
